initial commit to have all converter tests leverage identical testing logic
