+++ converter/internal/promtailconvert/testdata/bad_config.diags
