+++ converter/internal/promtailconvert/testdata/kubernetes.diags
