+++ converter/internal/promtailconvert/testdata/unsupported.diags
