+++ converter/internal/staticconvert/testdata/prom_scrape.diags
